backend/requirements.txt
numpy<2
fastapi
uvicorn[standard]
spacy==3.7.5
pdfminer.six
python-docx
pandas
scikit-learn
shap
fairlearn
aif360
python-dotenv
python-multipart
pytest
httpx
reportlab
pytesseract
pdf2image
Pillow
huggingface_hub
SQLAlchemy>=2.0.0
pymysql
cryptography
